feat: add ON_CLOSE handler
