+++ src/controllers/index.ts
@@ -1,2 +1,2 @@
-export * from './add-client.ctrl';
+export * from './client.ctrl';
 export * from './validate-command.ctrl';
