+++ src/ws_server/handle-dialog.ts
@@ -1,4 +1,4 @@
-import { addClientCtrl, validateCommandCtrl } from '../controllers';
+import { addClientCtrl, removeClientCtrl, validateCommandCtrl } from '../controllers';
 import { CommandType, Message } from '../models';
 import { sendMessages } from './send-messages';
 
@@ -16,6 +16,11 @@ export const handleDialog = (fromMessage: Message): void => {
   switch (fromMessage.command.type) {
     case CommandType.ON_CONNECTION: {
       const message = addClientCtrl(fromMessage);
+      toMessages.push(message);
+      break;
+    }
+    case CommandType.ON_CLOSE: {
+      const message = removeClientCtrl(fromMessage);
       toMessages.push(message);
       break;
     }
